requirements.txt
authlib==0.15.5
cachecontrol==0.12.10; python_version >= "3.6"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0")
certifi==2021.10.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.11; python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
click==8.0.3; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
cryptography==36.0.1; python_version >= "3.6"
dateparser==1.1.0; python_version >= "3.5"
firebase-admin==5.2.0; python_version >= "3.6"
flask==2.0.2; python_version >= "3.6"
google-api-core==2.5.0; platform_python_implementation != "PyPy" and python_version >= "3.7"
google-api-python-client==2.36.0; python_version >= "3.6"
google-auth-httplib2==0.1.0; python_version >= "3.6"
google-auth==2.6.0; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0"
google-cloud-core==2.2.2; platform_python_implementation != "PyPy" and python_version >= "3.7"
google-cloud-firestore==2.3.4; platform_python_implementation != "PyPy" and python_version >= "3.6"
google-cloud-storage==2.1.0; python_version >= "3.7"
google-crc32c==1.3.0; python_version >= "3.7"
google-resumable-media==2.1.0; python_version >= "3.7"
googleapis-common-protos==1.54.0; platform_python_implementation != "PyPy" and python_version >= "3.6"
grpcio-status==1.43.0; platform_python_implementation != "PyPy" and python_version >= "3.6"
grpcio==1.43.0; platform_python_implementation != "PyPy" and python_version >= "3.6"
httplib2==0.20.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
idna==3.3; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
itsdangerous==2.0.1; python_version >= "3.6"
jinja2==3.0.3; python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.6"
msgpack==1.0.3; python_version >= "3.6"
oauthlib==3.2.0; python_version >= "3.6"
packaging==21.3; platform_python_implementation != "PyPy" and python_version >= "3.6"
proto-plus==1.19.9; platform_python_implementation != "PyPy" and python_version >= "3.6"
protobuf==3.19.4; platform_python_implementation != "PyPy" and python_version >= "3.7"
pyasn1-modules==0.2.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0"
pyasn1==0.4.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0") or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0")
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyparsing==3.0.7; python_version >= "3.6" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.4.0" and python_version >= "3.6" and platform_python_implementation != "PyPy"
python-dateutil==2.8.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
pytz-deprecation-shim==0.1.0.post0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pytz==2021.3
pyyaml==6.0; python_version >= "3.6"
regex==2022.1.18; python_version >= "3.5"
requests-oauthlib==1.3.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
returns==0.17.0; python_version >= "3.7" and python_version < "4.0"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or platform_python_implementation != "PyPy" and python_version >= "3.7" and python_full_version >= "3.6.0")
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version >= "3.7" and platform_python_implementation != "PyPy"
typing-extensions==3.10.0.2
tzdata==2021.5; platform_system == "Windows" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
tzlocal==4.1; python_version >= "3.6"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.8; python_version >= "3.7" and python_full_version < "3.0.0" and platform_python_implementation != "PyPy" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7" and platform_python_implementation != "PyPy"
werkzeug==2.0.2; python_version >= "3.6"
